requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
cloudflare==2.19.2
cloudflareddns==0.0.8
filelock==3.13.3
idna==3.6
jsonlines==4.0.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
requests-file==2.0.0
tldextract==5.1.2
urllib3==2.2.1
